Move gps helper functions to iSToolsGNSS.py
