+++ python/logInspector/logPlotter.py
@@ -36,7 +36,7 @@ sys.path.append(os.path.normpath(file_path + '/..'))
 sys.path.append(os.path.normpath(file_path + '/../math/src'))
 
 from logReader import Log
-from pylib.ISToolsDataSorted import refLla, getTimeFromGpsTowMs, getTimeFromGpsTow, setGpsWeek, getTimeFromGTime, setShowUtcTime
+from pylib.ISToolsGNSS import refLla, getTimeFromGpsTowMs, getTimeFromGpsTow, setGpsWeek, getTimeFromGTime, setShowUtcTime
 from pylib.data_sets import *
 from inertialsense_math.pose import quat2euler, lla2ned, rotmat_ecef2ned, quatRot, quatConjRot, quat_ecef2ned
 import datetime
